ai-services/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
scikit-learn==1.4.0
numpy==1.26.3
pandas==2.1.4
openai==1.10.0
httpx==0.26.0
python-dotenv==1.0.0
redis==5.0.1
joblib==1.3.2
pytest==7.4.4
pytest-asyncio==0.23.3
